CTCP-2345: Update dependencies and prepare for secure-data-exchange-proxy

+++ app/models/sdes/SdesFile.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 HM Revenue & Customs
+ * Copyright 2023 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,10 +17,10 @@
 package models.sdes
 
 case class SdesFile(
-  recipientOrSender: String,
+  recipientOrSender: String, // SRN
   name: String,
   location: String,
   checksum: SdesChecksum,
-  size: Int,
-  properties: Seq[SdesProperties] // ?? message sender ??
+  size: Long,
+  properties: Seq[SdesProperties] // x-conversation-id
 )
